Moved schedulers to other package. Updated Note ids field type. Renamed NotepadVM field.
